+++ tests/test_rwlock.c
@@ -191,7 +191,8 @@ void *thr_reader(void *_count)
 	unsigned long long *count = _count;
 
 	printf_verbose("thread_begin %s, thread id : %lx, tid %lu\n",
-			"reader", pthread_self(), (unsigned long)gettid());
+			"reader", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	set_affinity();
 
@@ -212,7 +213,8 @@ void *thr_reader(void *_count)
 
 	*count = URCU_TLS(nr_reads);
 	printf_verbose("thread_end %s, thread id : %lx, tid %lu\n",
-			"reader", pthread_self(), (unsigned long)gettid());
+			"reader", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	return ((void*)1);
 
 }
@@ -222,7 +224,8 @@ void *thr_writer(void *_count)
 	unsigned long long *count = _count;
 
 	printf_verbose("thread_begin %s, thread id : %lx, tid %lu\n",
-			"writer", pthread_self(), (unsigned long)gettid());
+			"writer", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	set_affinity();
 
@@ -246,7 +249,8 @@ void *thr_writer(void *_count)
 	}
 
 	printf_verbose("thread_end %s, thread id : %lx, tid %lu\n",
-			"writer", pthread_self(), (unsigned long)gettid());
+			"writer", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	*count = URCU_TLS(nr_writes);
 	return ((void*)2);
 }
@@ -352,7 +356,8 @@ int main(int argc, char **argv)
 	printf_verbose("Writer delay : %lu loops.\n", wdelay);
 	printf_verbose("Reader duration : %lu loops.\n", rduration);
 	printf_verbose("thread %-6s, thread id : %lx, tid %lu\n",
-			"main", pthread_self(), (unsigned long)gettid());
+			"main", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	tid_reader = malloc(sizeof(*tid_reader) * nr_readers);
 	tid_writer = malloc(sizeof(*tid_writer) * nr_writers);
