+++ tests/test_rwlock_timing.c
@@ -87,7 +87,8 @@ void *thr_reader(void *arg)
 	cycles_t time1, time2;
 
 	printf("thread_begin %s, thread id : %lx, tid %lu\n",
-			"reader", pthread_self(), (unsigned long)gettid());
+			"reader", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	sleep(2);
 
 	time1 = caa_get_cycles();
@@ -104,7 +105,8 @@ void *thr_reader(void *arg)
 
 	sleep(2);
 	printf("thread_end %s, thread id : %lx, tid %lu\n",
-			"reader", pthread_self(), (unsigned long)gettid());
+			"reader", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	return ((void*)1);
 
 }
@@ -115,7 +117,8 @@ void *thr_writer(void *arg)
 	cycles_t time1, time2;
 
 	printf("thread_begin %s, thread id : %lx, tid %lu\n",
-			"writer", pthread_self(), (unsigned long)gettid());
+			"writer", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	sleep(2);
 
 	for (i = 0; i < OUTER_WRITE_LOOP; i++) {
@@ -131,7 +134,8 @@ void *thr_writer(void *arg)
 	}
 
 	printf("thread_end %s, thread id : %lx, tid %lu\n",
-			"writer", pthread_self(), (unsigned long)gettid());
+			"writer", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 	return ((void*)2);
 }
 
@@ -157,7 +161,8 @@ int main(int argc, char **argv)
 	tid_writer = malloc(sizeof(*tid_writer) * num_write);
 
 	printf("thread %-6s, thread id : %lx, tid %lu\n",
-			"main", pthread_self(), (unsigned long)gettid());
+			"main", (unsigned long) pthread_self(),
+			(unsigned long) gettid());
 
 	for (i = 0; i < NR_READ; i++) {
 		err = pthread_create(&tid_reader[i], NULL, thr_reader,
